api/UwagaImpreza/requirements.txt

# -- for backend and database
Django==3.1.4

# -- for rest api creation
djangorestframework==3.12.2

# -- for PostgreSQL integration
psycopg2-binary==2.8.6

# -- cors headers
django-cors-headers==3.5.0

# -- for image storage
Pillow==8.0.1

# -- for user authentication
djangorestframework-simplejwt==4.6.0

